+++ src/main/java/connectripbe/connectrip_be/member/service/MemberService.java
@@ -8,7 +8,6 @@ import connectripbe.connectrip_be.member.dto.MemberHeaderInfoDto;
 import connectripbe.connectrip_be.member.dto.ProfileDto;
 import connectripbe.connectrip_be.member.dto.ProfileUpdateRequestDto;
 import connectripbe.connectrip_be.member.dto.TokenAndHeaderInfoDto;
-import connectripbe.connectrip_be.review.dto.AccompanyReviewListResponse;
 
 public interface MemberService {
 
@@ -21,8 +20,6 @@ public interface MemberService {
     TokenAndHeaderInfoDto getFirstUpdateMemberResponse(String tempTokenCookie, FirstUpdateMemberRequest request);
 
     ProfileDto getProfile(Long memberId);  // 프로필 조회 (최신 3개 리뷰 포함)
-
-    AccompanyReviewListResponse getAllReviews(Long memberId);  // 모든 리뷰 조회
 
     String calculateAgeGroup(int age);  // 나이대 계산 메서드
